First Exercise
- Moved "products" array to its own file (data.js)

- Completed "buy(id)" function

Note: In order to access the array from data.js I had to add the type="module" attribute to the index.html file. I believe, because of that I had to convert functions into modular functions by adding the line of code: "window.function = function"
