+++ index.html
@@ -273,7 +273,7 @@
         <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/js/bootstrap.bundle.min.js" integrity="sha384-ka7Sk0Gln4gmtz2MlQnikT1wXgYsOg+OMhuP+IlRH9sENBO0LRn5q+8nbTov4+1p" crossorigin="anonymous"></script>
 
         <!-- Core theme JS-->
-        <script src="js/shop.js"></script>
-		<script src="js/checkout.js"></script>
+        <script src="js/shop.js" type="module"></script>
+		<script src="js/checkout.js" type="module"></script>
     </body>
 </html>
