+++ js/data.js
@@ -1,0 +1,67 @@
+
+var products = [
+  {
+       id: 1,
+       name: 'cooking oil',
+       price: 10.5,
+       type: 'grocery',
+       offer: {
+           number: 3,
+           percent: 20
+       }
+   },
+   {
+       id: 2,
+       name: 'Pasta',
+       price: 6.25,
+       type: 'grocery'
+   },
+   {
+       id: 3,
+       name: 'Instant cupcake mixture',
+       price: 5,
+       type: 'grocery',
+       offer: {
+           number: 10,
+           percent: 30
+       }
+   },
+   {
+       id: 4,
+       name: 'All-in-one',
+       price: 260,
+       type: 'beauty'
+   },
+   {
+       id: 5,
+       name: 'Zero Make-up Kit',
+       price: 20.5,
+       type: 'beauty'
+   },
+   {
+       id: 6,
+       name: 'Lip Tints',
+       price: 12.75,
+       type: 'beauty'
+   },
+   {
+       id: 7,
+       name: 'Lawn Dress',
+       price: 15,
+       type: 'clothes'
+   },
+   {
+       id: 8,
+       name: 'Lawn-Chiffon Combo',
+       price: 19.99,
+       type: 'clothes'
+   },
+   {
+       id: 9,
+       name: 'Toddler Frock',
+       price: 9.99,
+       type: 'clothes'
+   }
+]
+
+export default products
